% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbat.R
\name{jatosify}
\alias{jatosify}
\title{Create a JATOS .jzip file}
\usage{
jatosify(
  study_title,
  html_file_list,
  JATOS_version,
  study_desc = "",
  study_comment = "",
  output_dir = "."
)
}
\arguments{
\item{study_title}{The title of the study. Used for filenames.}

\item{html_file_list}{A vector of HTML filenames to be used as JATOS components (order is preserved).}

\item{JATOS_version}{The version of the study (e.g., "3.9.0").}

\item{study_desc}{A short description of the study (optional).}

\item{study_comment}{Comments about the study (optional).}

\item{output_dir}{The output directory for the .jzip file (defaults to the current directory).}
}
\description{
Create a JATOS .jzip file
}
\examples{
# jatosify("exp01",c("ic.html","age_gender.html","task01.html"),"3.9")
}
