% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatos.R
\name{print.jatos_data}
\alias{print.jatos_data}
\title{Print a summary of the object returned by readJatos()}
\usage{
\method{print}{jatos_data}(x, ...)
}
\arguments{
\item{x}{object of class \code{jatos_data}}

\item{...}{ignored}
}
\description{
Print a summary of the object returned by readJatos()
}
