% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{researchOut}
\alias{researchOut}
\title{upload Japanese e-labnotebook to GitHub}
\usage{
researchOut(rc = TRUE)
}
\arguments{
\item{rc}{If you are using Research Compendium of senshuRmd,
you can create a e-labnotebook file in the "labnote" directory from the current directory.
In that case, please set rc to TURE.}
}
\description{
upload Japanese e-labnotebook to GitHub
}
\examples{
# researchOut()
}
