% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highLoad.R
\name{runHighLoad}
\alias{runHighLoad}
\title{run high load computing and psuh to GitHub}
\usage{
runHighLoad(r_file)
}
\arguments{
\item{r_file}{set r file to run}
}
\description{
run high load computing and psuh to GitHub
}
\examples{
# runHighLoad(r_file)
}
