% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbat.R
\name{set_phaser}
\alias{set_phaser}
\title{Set template files for Phaser3}
\usage{
set_phaser(game_name = "game_name", phaser_version = "3.80.0", use_rc = TRUE)
}
\arguments{
\item{game_name}{name of task}

\item{phaser_version}{version of phaser}

\item{use_rc}{If you don"t use the RC, set FALSE.}
}
\description{
Set template files for Phaser3
}
\examples{
# set_phaser("game1","3.80.0",FALSE)
}
