% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highLoad.R
\name{setGitSlack}
\alias{setGitSlack}
\title{run high load computing and psuh to GitHub}
\usage{
setGitSlack(
  git_name,
  git_email,
  slack_token = "FALSE",
  slack_channel = "#general",
  slack_username = "R"
)
}
\arguments{
\item{git_name}{set user name of GitHub}

\item{git_email}{set e-mail of GitHub}

\item{slack_token}{set slack API token}

\item{slack_channel}{set channel name of slack}

\item{slack_username}{set user name of slack}
}
\description{
run high load computing and psuh to GitHub
}
\examples{
# setGitSlack(user_name_github, email_github, slack_token, slack_channel, slack_username)
}
