% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{up2osf}
\alias{up2osf}
\title{upload Japanese e-labnotebook to OSF}
\usage{
up2osf(eln_osf = TRUE, rc_osf = TRUE)
}
\arguments{
\item{eln_osf}{URL of pdf directory in OSF}

\item{rc_osf}{If you are using Research Compendium of senshuRmd,
you can create a e-labnotebook file in the "labnote" directory from the current directory.
In that case, please set rc to TURE.}
}
\description{
upload Japanese e-labnotebook to OSF
}
